xai/requirements.txt
torch>=1.10.0
numpy>=1.22.0
matplotlib>=3.5.0
librosa>=0.9.0
scipy>=1.7.0
transformers>=4.20.0
opencv-python>=4.5.0
ultralytics>=8.0.0
Pillow>=9.0.0 